Async Python APIs/python-mqtt-docker/requirements.txt
paho-mqtt==1.6.1
Flask==2.2.2
Werkzeug==2.2.2